Renamed header and footer
