+++ application/src/main/java/de/starwit/SaeVisualizerApplication.java
@@ -4,10 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SAEVisualizerApplication {
+public class SaeVisualizerApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(SAEVisualizerApplication.class, args);
+		SpringApplication.run(SaeVisualizerApplication.class, args);
 	}
 
 }
